+++ src/main/java/tis/project/lion/postproject/service/PostService.java
@@ -8,14 +8,6 @@ import tis.project.lion.postproject.domain.PostDto;
 import java.util.List;
 
 public interface PostService {
-    //board CRUD;
-    public void createBoard(BoardDto boardDto);
-
-    public Board findBoardOne(BoardDto boardDto);
-
-    public Board updateBoard(BoardDto boardDto);
-
-    public void deleteBoard(BoardDto boardDto);
     //post CRUD;
 
     public void createPost(PostDto postDto);
@@ -24,7 +16,7 @@ public interface PostService {
 
     public Post findPostOne(PostDto postDto);
 
-    public Post updatePost(PostDto postDto);
+    public Post editPost(PostDto postDto);
 
     public void deletePost(PostDto postDto);
 }
