--- src/main/java/tis/project/lion/postproject/web/OutputBoardForm.java
@@ -1,4 +1,0 @@
-package tis.project.lion.postproject.web;
-
-public class OutputBoardForm {
-}
