+++ src/main/java/tis/project/lion/postproject/web/controller/BoardController.java
@@ -1,5 +1,6 @@
-package tis.project.lion.postproject.web;
+package tis.project.lion.postproject.web.controller;
 
+import tis.project.lion.postproject.web.form.InputBoardForm;
 import tis.project.lion.postproject.web.message.ResponseMessage;
 
 public interface BoardController {
@@ -8,6 +9,6 @@ public interface BoardController {
 //    public String createForm();
     public ResponseMessage createBoard(InputBoardForm boardForm);
 //    public String updateForm(Long board_id);
-    public ResponseMessage updateBoard(InputBoardForm boardForm);
+    public ResponseMessage editBoard(InputBoardForm boardForm);
     public ResponseMessage deleteBoard(Long board_id);
 }
