+++ src/main/java/tis/project/lion/postproject/web/form/InputBoardForm.java
@@ -1,0 +1,4 @@
+package tis.project.lion.postproject.web.form;
+
+public class InputBoardForm {
+}
