+++ src/main/java/tis/project/lion/postproject/web/form/InputPostForm.java
@@ -1,0 +1,4 @@
+package tis.project.lion.postproject.web.form;
+
+public class InputPostForm {
+}
