+++ src/main/java/tis/project/lion/postproject/web/form/OutputBoardForm.java
@@ -1,0 +1,4 @@
+package tis.project.lion.postproject.web.form;
+
+public class OutputBoardForm {
+}
